Define top tags resolver (#15)
This PR defines a Graphene resolver for fetching the top N tags as specified in the GraphQL query. By default, N is 10. The "top" is defined as the tag with the most associated beers. This will most likely change in the future and become a more complicated calculation in the future, but is a good starting point.

I also:

    Consolidated fixtures into 1 JSON file, so we can run with 1 command
    Create a Tag model
    Regenereated initial migration