Boilerplate setup using Cline (#22)
This PR performs some boilerplate setup using Cline. In the front-end, it sets up a basic home page that displays a tiled list of beers alongside some information about each. In the back-end, this adds the GraphQL queries for supporting the front-end's home page. It also sets up a script to generate some dummy beer data to be displayed.