Resolve some cross environment infra naming conflicts (#41)
This PR fixes a few naming conflicts for various CDK constructs that set the name to a specific value. Since, for now, I plan to deploy both staging and prod to the same account, both stacks couldn't be deployed due to these conflicts. To fix, I added the env name to those names.