Migrate Django app to Railway instead of AWS (#68)
This PR migrates the Django back-end to use Railway, a service that provides easier server/DB management, rather than AWS. The main motivation for this is that this is a hobby project, so I'd like to 1) be able to focus on building new features rather than managing deployment and 2) the cost of AWS was quite large with the previous set up ($50-80 / month). I explored simplifying the architecture a bit while staying on AWS, but deploying to Railway was simpler.